pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lennyverse-ingestion"
version = "0.1.0"
description = "Local manual ingestion pipeline for LennyVerse foundation data."
requires-python = ">=3.11"
dependencies = [
    "neo4j>=5.28.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-ollama>=0.3.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.28.3",
    "pydantic>=2.12.5",
    "fastapi>=0.122.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.8.0", "httpx>=0.27.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["ingest", "backend"]
